♻️ refacto: added request method check in handlers

+++ handlers.go
@@ -9,6 +9,11 @@ import (
 
 func pingHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodGet {
+			http.Error(w, "Invalid method", http.StatusMethodNotAllowed)
+			return
+		}
+
 		marshal, err := json.Marshal("Pong!")
 		if err != nil {
 			return
@@ -24,6 +29,11 @@ func pingHandler() http.HandlerFunc {
 
 func encodeHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodPost {
+			http.Error(w, "Invalid method", http.StatusMethodNotAllowed)
+			return
+		}
+
 		request := parseRequest(r.Body, w)
 		link := request.Link
 		expiration := request.Expiration
@@ -33,6 +43,11 @@ func encodeHandler() http.HandlerFunc {
 
 func decodeHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodGet {
+			http.Error(w, "Invalid method", http.StatusMethodNotAllowed)
+			return
+		}
+
 		vars := strings.Split(r.RequestURI, "/")
 		code := vars[len(vars)-1]
 		handleDecodeRequest(w, r, code)
@@ -41,6 +56,11 @@ func decodeHandler() http.HandlerFunc {
 
 func purgeHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodGet {
+			http.Error(w, "Invalid method", http.StatusMethodNotAllowed)
+			return
+		}
+
 		rds := database.CreateClient(1)
 		err := rds.FlushAll(r.Context()).Err()
 		if err != nil {
